move RepoSearch from run package to jobutil
