+++ internal/search/job/jobutil/repos.go
@@ -1,4 +1,4 @@
-package run
+package jobutil
 
 import (
 	"context"
